+++ configs/nvim/init.lua
@@ -1,5 +1,0 @@
-local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
-vim.opt.rtp:prepend(lazypath)
-
-require("vim-options")
-require("lazy").setup("plugins")