+++ CHANGELOG.md
@@ -4,4 +4,4 @@
 - Added adw-gtk3 theme and configured a default font throughout using it
 - Configured the wallpaper to stay stationary during overview
 - Set default theme as Catpuccin (Changing colors based on wallpaper filled my memory idk why)
-
+- Configured nvim.nix